requirements.txt
# ============================================================
# 🚀 CORE API BACKEND
# ============================================================
Flask==3.0.3
fastapi==0.115.4
uvicorn==0.32.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3

# ============================================================
# 🧠 EMBEDDING / RAG ENGINE
# ============================================================
torch==2.8.0
sentence-transformers==3.2.1
transformers==4.46.2
langchain==0.3.7
langchain-community==0.3.7
qdrant-client==1.13.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.2

# ============================================================
# 💾 DATABASE & CONNECTORS
# ============================================================
mysql-connector-python==9.0.0
huggingface-hub==0.25.1
pydantic==2.9.2

# ============================================================
# 📄 OCR / DOCUMENT PROCESSING
# ============================================================
paddleocr==2.8.1
paddlepaddle==3.1.1
pymupdf==1.24.9
python-docx==1.1.2
openpyxl==3.1.5

# ============================================================
# 🧰 UTILITIES & HELPERS
# ============================================================
tqdm==4.67.1
filelock==3.16.1
typing-extensions==4.12.2
